# docs/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "NodeHub -  API 文档\n\n这是 NodeHub 的 API 文档\n\n## 认证\n大多数接口需要使用 JWT 令牌进行认证。\n认证时，请在 Authorization 头中包含 JWT 令牌：\n`Authorization: Bearer <your-jwt-token>`\n\n## 错误响应\n所有错误响应都遵循统一格式，包含 code、message 和 error 字段。\n\n## 成功响应\n所有成功响应都遵循统一格式，包含 code、message 和 data 字段。",
    "title": "NodeHub API",
    "contact": {
      "name": "NodeHub API 支持",
      "email": "support@nodehub.com"
    },
    "license": {
      "name": "GPL-3.0",
      "url": "https://opensource.org/license/gpl-3-0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/share/sub/{token}": {
      "get": {
        "description": "获取订阅内容 带规则的订阅",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/plain"
        ],
        "tags": [
          "分享"
        ],
        "summary": "获取订阅内容 带规则的订阅",
        "parameters": [
          {
            "type": "string",
            "description": "分享token",
            "name": "token",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功，内容为yaml/plain格式",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          }
        }
      }
    },
...
    "/api/v1/sub": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "订阅"
        ],
        "summary": "获取订阅链接",
        "parameters": [
          {
            "type": "integer",
            "description": "链接ID",
            "name": "id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_sub.Response"
                      }
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "401": {
            "description": "未授权",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "创建单个订阅链接",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "订阅"
        ],
        "summary": "创建订阅链接",
        "parameters": [
          {
            "description": "创建订阅链接请求",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_sub.Request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "创建成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_sub.Response"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "401": {
            "description": "未授权",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          }
        }
      }
    },
    "/api/v1/sub/batch": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "批量创建多个订阅链接",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "订阅"
        ],
        "summary": "批量创建订阅链接",
        "parameters": [
          {
            "description": "批量创建订阅链接请求",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_sub.Request"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "创建成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_sub.Response"
                      }
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "401": {
            "description": "未授权",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          }
        }
      }
    },
    "/api/v1/sub/refresh/{id}": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "根据ID手动刷新单个订阅",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "订阅"
        ],
        "summary": "手动刷新订阅",
        "parameters": [
          {
            "type": "integer",
            "description": "订阅链接ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "刷新成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_sub.Result"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "401": {
            "description": "未授权",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "404": {
            "description": "订阅链接不存在",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          }
        }
      }
    },
    "/api/v1/sub/{id}": {
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "根据请求体中的ID更新订阅链接信息",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "订阅"
        ],
        "summary": "更新订阅链接",
        "parameters": [
          {
            "type": "integer",
            "description": "订阅链接ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "更新订阅链接请求",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_sub.Request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_sub.Response"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "401": {
            "description": "未授权",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "404": {
            "description": "订阅链接不存在",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "根据ID删除单个订阅链接",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "订阅"
        ],
        "summary": "删除订阅链接",
        "parameters": [
          {
            "type": "integer",
            "description": "订阅链接ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "401": {
            "description": "未授权",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "404": {
            "description": "订阅链接不存在",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          }
        }
      }
    },
...
    "/api/v1/sub/log": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "订阅"
        ],
        "summary": "获取订阅运行日志",
        "description": "获取订阅拉取/转换日志，包含运行记录与事件",
        "parameters": [
          {
            "type": "integer",
            "description": "订阅ID",
            "name": "sub_id",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "description": "返回条数",
            "name": "limit",
            "in": "query"
          },
          {
            "name": "include_events",
            "in": "query",
            "description": "是否包含历史 events",
            "required": false,
            "type": "boolean",
            "default": false
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_sub_RunLogResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "401": {
            "description": "未授权",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "github_com_evilCYH_nodehub_internal_models_node.SimpleInfo": {
      "type": "object",
      "properties": {
        "count": {
          "type": "integer"
        },
        "delay": {
          "type": "integer"
        },
        "risk": {
          "type": "integer"
        },
        "speed_down": {
          "type": "integer"
        },
        "speed_up": {
          "type": "integer"
        }
      }
    },
...
    "github_com_evilCYH_nodehub_internal_models_sub.Config": {
      "type": "object",
      "properties": {
        "protocol_filter": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "protocol_filter_enable": {
          "type": "boolean"
        },
        "protocol_filter_mode": {
          "type": "boolean"
        },
        "proxy": {
          "type": "boolean"
        },
        "timeout": {
          "type": "integer"
        },
        "url": {
          "type": "string"
        }
      }
    },
    "github_com_evilCYH_nodehub_internal_models_sub.Request": {
      "type": "object",
      "properties": {
        "config": {
          "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_sub.Config"
        },
        "cron_expr": {
          "type": "string",
          "example": "0 0 * * *"
        },
        "enable": {
          "type": "boolean"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "github_com_evilCYH_nodehub_internal_models_sub.Response": {
      "type": "object",
      "properties": {
        "config": {
          "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_sub.Config"
        },
        "created_at": {
          "type": "string"
        },
        "cron_expr": {
          "type": "string"
        },
        "enable": {
          "type": "boolean"
        },
        "id": {
          "type": "integer"
        },
        "info": {
          "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_node.SimpleInfo"
        },
        "name": {
          "type": "string"
        },
        "result": {
          "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_sub.Result"
        },
        "status": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        }
      }
    },
    "github_com_evilCYH_nodehub_internal_models_sub.Result": {
      "type": "object",
      "properties": {
        "duration": {
          "type": "integer"
        },
        "fail": {
          "type": "integer"
        },
        "last_run": {
          "type": "string"
        },
        "msg": {
          "type": "string"
        },
        "node_null_count": {
          "type": "integer"
        },
        "raw_count": {
          "type": "integer"
        },
        "success": {
          "type": "integer"
        }
      }
    },
...
    "github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 200
        },
        "data": {},
        "message": {
          "type": "string",
          "example": "success"
        }
      }
    },
    "github_com_evilCYH_nodehub_internal_models_sub_RunLog": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "sub_id": {
          "type": "integer"
        },
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "raw_count": {
          "type": "integer"
        },
        "accepted": {
          "type": "integer"
        },
        "created_at": {
          "type": "string"
        },
        "duration_ms": {
          "type": "integer"
        }
      }
    },
    "github_com_evilCYH_nodehub_internal_models_sub_RunEvent": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "run_id": {
          "type": "integer",
          "format": "int64"
        },
        "sub_id": {
          "type": "integer"
        },
        "step": {
          "type": "string"
        },
        "level": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "run_time": {
          "type": "string"
        }
      }
    },
    "github_com_evilCYH_nodehub_internal_models_sub_RunLogResponse": {
      "type": "object",
      "properties": {
        "runs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_sub_RunLog"
          }
        },
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_sub_RunEvent"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "类型为 \"Bearer\"，后跟空格和 JWT 令牌。",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}